Address coder PR handoff reviews
