+++ crates/tandem-server/src/http/coder.rs
@@ -8,7 +8,9 @@ include!("coder_parts/part01.rs");
 include!("coder_parts/part02.rs");
 include!("coder_parts/part03.rs");
 include!("coder_parts/part04.rs");
+include!("coder_parts/part10.rs");
 include!("coder_parts/part05.rs");
 include!("coder_parts/part06.rs");
+include!("coder_parts/part09.rs");
 include!("coder_parts/part07.rs");
 include!("coder_parts/part08.rs");
