+++ crates/tandem-server/src/http/tests/bug_monitor.rs
@@ -6,3 +6,4 @@ use tokio::sync::RwLock;
 include!("bug_monitor_parts/part01.rs");
 include!("bug_monitor_parts/part03.rs");
 include!("bug_monitor_parts/part02.rs");
+include!("bug_monitor_parts/part04.rs");
